+++ src/Rhetos.MsSqlEf6/MsSqlUtility.cs
@@ -455,7 +455,7 @@ namespace Rhetos.Utilities
         /// Returns empty string if the string value is null.
         /// This function is used for compatibility between MsSql and Oracle string behavior.
         /// </summary>
-        public string EmptyNullString(DbDataReader dataReader, int column)
+        public string ReadEmptyNullString(DbDataReader dataReader, int column)
         {
             return dataReader.GetString(column) ?? "";
         }